Revamped toggle button to be put on a new Navigation bar that loads on page.html load
A critical error occured with my previous code in which due to the toggle button being integrated on the left.html, it would be customisable by the adaptive UI which is not what we want. Firebase also overwrites the left.html which can lead to errors in fetching Page.js.

In return I revamped it to be on a new item being the nav bar. Which is not customisable and affected by the adaptive UI element so that the option to hide and show the right side will always be available